feat: moved assertions of nickname and password to commons
